refactor: :bulb:  rename modules
